+++ NEWS.md
@@ -1,3 +1,5 @@
+# furrr (development version)
+
 # furrr 0.2.1
 
 * Updated documentation examples to explicitly set the seed on the workers